+++ apps/fluent-tester/src/E2E/Persona/pages/PersonaTestPage.win.ts
@@ -1,4 +1,4 @@
-import { PERSONA_TESTPAGE } from '../../../RNTester/TestComponents/Persona/consts';
+import { PERSONA_TESTPAGE } from '../../../FluentTester/TestComponents/Persona/consts';
 import { BasePage, By } from '../../common/BasePage';
 
 class PersonaTestPage extends BasePage {