+++ packages/framework/theming-ramp/src/index.ts
@@ -1,6 +1,13 @@
-export * from './Color.types';
-export * from './Color';
 export * from './SettingsWorker';
-export * from './Theme.types';
 export * from './Theme';
-export * from './Typography.types';
+export {
+  ColorValue,
+  OfficePalette,
+  Palette as IPalette,
+  PartialPalette as IPartialPalette,
+  PartialTheme as IPartialTheme,
+  Theme as ITheme,
+  ThemeColorDefinition as IThemeColorDefinition,
+  ThemeContext,
+  Typography as ITypography,
+} from '@fluentui-react-native/theme-types';