+++ packages/theming/default-theme/src/index.ts
@@ -1,0 +1,2 @@
+export { defaultFluentTheme, defaultFluentDarkTheme } from './defaultTheme';
+export * from './createDefaultTheme';