+++ scripts/babel.config.js
@@ -1,5 +1,3 @@
-const path = require('path');
-
 module.exports = {
   presets: [
     ['@babel/preset-env', { targets: { node: 'current' } }],