+++ packages/components/Checkbox/CHANGELOG.md
@@ -1,8 +1,33 @@
 # Change Log - @fluentui-react-native/checkbox
 
-This log was last generated on Mon, 17 Aug 2020 22:08:34 GMT and should not be manually modified.
+This log was last generated on Mon, 28 Sep 2020 20:32:21 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.5.10
+
+Mon, 28 Sep 2020 20:32:21 GMT
+
+### Patches
+
+- move test pages (aliciakds88@gmail.com)
+
+## 0.5.8
+
+Fri, 25 Sep 2020 19:21:43 GMT
+
+### Patches
+
+- Update react-native-win32 versions - enable logbox (acoates-ms@noreply.github.com)
+
+## 0.5.5
+
+Tue, 22 Sep 2020 04:57:02 GMT
+
+### Patches
+
+- Update a11y properties (krsiler@microsoft.com)
+- manual version bump on Checkbox (taamireh@microsoft.com)
 
 ## 0.5.2