+++ packages/framework/foundation-composable/CHANGELOG.md
@@ -1,8 +1,24 @@
 # Change Log - @uifabricshared/foundation-composable
 
-This log was last generated on Thu, 20 Aug 2020 00:14:01 GMT and should not be manually modified.
+This log was last generated on Fri, 25 Sep 2020 19:21:43 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.6.75
+
+Fri, 25 Sep 2020 19:21:43 GMT
+
+### Patches
+
+- Update react-native-win32 versions - enable logbox (acoates-ms@noreply.github.com)
+
+## 0.6.74
+
+Wed, 23 Sep 2020 18:31:48 GMT
+
+### Patches
+
+- start publishing src to fix customer source maps (jasonmo@microsoft.com)
 
 ## 0.6.73