+++ packages/libraries/core/CHANGELOG.md
@@ -1,8 +1,48 @@
 # Change Log - @fluentui/react-native
 
-This log was last generated on Wed, 09 Sep 2020 21:36:19 GMT and should not be manually modified.
+This log was last generated on Mon, 28 Sep 2020 20:32:21 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.18.0
+
+Mon, 28 Sep 2020 20:32:21 GMT
+
+### Minor changes
+
+- add more host colors (aliciakds88@gmail.com)
+
+## 0.17.0
+
+Mon, 28 Sep 2020 19:03:01 GMT
+
+### Minor changes
+
+- Adding FocusZone component (safreibe@microsoft.com)
+
+## 0.16.7
+
+Fri, 25 Sep 2020 19:21:43 GMT
+
+### Patches
+
+- Update react-native-win32 versions - enable logbox (acoates-ms@noreply.github.com)
+
+## 0.16.4
+
+Wed, 23 Sep 2020 18:31:48 GMT
+
+### Patches
+
+- start publishing src to fix customer source maps (jasonmo@microsoft.com)
+
+## 0.16.3
+
+Tue, 22 Sep 2020 21:42:19 GMT
+
+### Patches
+
+- Updated Shimmer to not have a module name (saadnajmi2@gmail.com)
 
 ## 0.16.1